requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
Django==5.2.5
djangorestframework==3.16.1
et_xmlfile==2.0.0
fonttools==4.59.1
html5lib==1.1
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tzdata==2025.2
weasyprint==60.2
webencodings==0.5.1
zopfli==0.2.3.post1
